requirements.txt
# Core dependencies
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
psycopg2-binary==2.9.9

# Shopify
shopify-python-api==12.0.0

# Web scraping (for module 2)
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Data processing
pandas==2.1.3

# OpenAI API
openai==1.3.7

# Email
smtplib # Built-in

# Logging and utilities
colorlog==6.7.0